// src/app/models/course.model.ts
export interface Course{
    courseId: string;
    courseName: string;
    technology: string;
    startDate: any;
    endDate: any;
    launchUrl: string;
    details: string;
}

export interface CourseTable{
    courseId: string;
    courseName: string;
    technology: string;
    startDate: any;
    endDate: any;
    launchUrl: string;
    details: string;
    delete: string;
}